+++ week01/02_odd_or_even.go
