+++ week01/03_array_info
