+++ week03/hello.c
@@ -1,0 +1,5 @@
+#include <stdio.h>
+int main() {
+    printf("Hello, world!\n");
+    return 0;
+}
