+++ week03/hello.go
@@ -1,0 +1,11 @@
+package main
+
+// #include <stdio.h>
+// void printHelloWorld() {
+//     printf("Hello, world!\n");
+// }
+import "C"
+
+func main() {
+	C.printHelloWorld()
+}
